echorooms/requirements.txt
streamlit
pymongo
python-dotenv
transformers
torch
librosa
langchain
faiss-cpu
openai
tqdm
